+++ ui_automation_tests/features/goods.feature
@@ -29,20 +29,6 @@ Feature: I want to edit and remove goods on the goods list
     Then I see good information
     And I see the good is in a query
 
-  @LT_886_add_new_good_to_app @regression
-  Scenario: Add a new good directly to a standard application
-    Given I go to exporter homepage and choose Test Org
-    When I create a standard application
-    And I click on the "goods" section
-    Then I see there are no goods on the application
-    When I click Add a new good
-    And I add a new good with description "New good for application" part number "P123" controlled "Yes" control code "ML8a25b" and graded "yes"
-    And I add the goods grading with prefix "abc" grading "nato_restricted" suffix "def" issuing authority "NATO" reference "12345" Date of issue "10-05-2015"
-    And I confirm I can upload a document
-    And I attach a document to the good with description "Test good spec 01"
-    And I enter details for the new good on an application with value "99.98", quantity "13" and unit of measurement "Kilograms" and I click Continue
-    Then A new good has been added to the application
-
   @LT_1262_add_good_without_document @regression
   Scenario: Add a new good without a document for a valid reason
     Given I go to exporter homepage and choose Test Org
