+++ ui_automation_tests/features/hmrc.feature
@@ -12,7 +12,7 @@ Feature: I want to be able to perform actions as a HMRC user
     And I click on the "end_user" section
     And I add an end user of sub_type: "government", name: "Mr Smith", website: "https://www.smith.com", address: "London" and country "Ukraine"
     And I upload a file "file_for_doc_upload_test_1.txt"
-    And I wait for document to upload
+    And I wait for download link
     And I click the back link
     And I click on the "goods" section
     When I add a goods type with description "M4"
