+++ ui_automation_tests/features/members.feature
@@ -11,10 +11,13 @@ Feature: I want to manage my organisation's members
     And I add a member to the organisation
     And I select the member that was just added
     # Edit the user
-    And I deactivate them, then the member is deactivated
-    And I reactivate them, then the member is reactivated
-    And I change what sites they're assigned to
-    Then I change their role
+    And I deactivate them
+    Then the member is deactivated
+    When I reactivate them
+    Then the member is reactivated
+    When I change what sites they're assigned to
+    And I change their role to Super User
+    Then role is changed
 
   @deactivate_oneself @regression
   Scenario: Deactivate oneself
@@ -30,15 +33,17 @@ Feature: I want to manage my organisation's members
     And filter status has been changed to "Active"
     Then I see the new member
     When I select the member that was just added
-    And I deactivate them, then the member is deactivated
-    And I click on the manage my organisation link
-    When I show filters
+    And I deactivate them
+    Then the member is deactivated
+    When I click on the manage my organisation link
+    And I show filters
     And filter status has been changed to "Active"
     Then I do not see the new member
     When filter status has been changed to "All"
     Then I see the new member
     When I select the member that was just added
-    And I reactivate them, then the member is reactivated
+    And I reactivate them
+    Then the member is reactivated
     And I click on the manage my organisation link
     And I show filters
     And filter status has been changed to "Active"
