+++ ui_automation_tests/features/submit_mod_application.feature
@@ -22,17 +22,17 @@ Feature: I want to create MOD Licence Applications
     When I click on the "end_user" section
     And I add an end user of sub_type: "government", name: "Mr Smith", website: "https://www.smith.com", address: "London" and country "Ukraine"
     And I upload a file "file_for_doc_upload_test_1.txt"
-    Then wait for download link
-    When I click the back link
+    And I wait for download link
+    And I click the back link
     And I click on the "consignee" section
     And I add an end user of sub_type: "government", name: "Mr Smith", website: "https://www.smith.com", address: "London" and country "Ukraine"
     And I upload a file "file_for_doc_upload_test_1.txt"
-    Then wait for download link
-    When I click the back link
+    And I wait for download link
+    And I click the back link
     And I submit the application
     Then application is submitted
     When I go to exporter homepage
-    And I click applications
+    And I click on applications
     Then I see submitted application
 
   @LT_1980_MOD_clearance @setup @smoke
@@ -50,12 +50,12 @@ Feature: I want to create MOD Licence Applications
     And I add an end user with clearance of sub_type: "government", name: "Mr Smith", website: "https://www.smith.com", clearance: "uk_unclassified", address: "London" and country "Ukraine"
     And I click continue
     And I upload a file "file_for_doc_upload_test_1.txt"
-    Then wait for download link
-    When I click the back link
+    And I wait for download link
+    And I click the back link
     And I submit the application
     Then application is submitted
     When I go to exporter homepage
-    And I click applications
+    And I click on applications
     Then I see submitted application
 
   @LT_1980_gifting_clearance @setup @smoke
@@ -70,10 +70,10 @@ Feature: I want to create MOD Licence Applications
     When I click on the "end_user" section
     And I add an end user of sub_type: "government", name: "Mr Smith", website: "https://www.smith.com", address: "London" and country "Ukraine"
     And I upload a file "file_for_doc_upload_test_1.txt"
-    Then wait for download link
-    When I click the back link
+    And I wait for download link
+    And I click the back link
     And I submit the application
     Then application is submitted
     When I go to exporter homepage
-    And I click applications
+    And I click on applications
     Then I see submitted application
