+++ ui_automation_tests/pages/exporter_hub_page.py
@@ -5,10 +5,8 @@ class ExporterHubPage(BasePage):
     APPLY_FOR_A_LICENCE_BTN = "a[href*='/apply-for-a-licence/']"
     APPLICATIONS_BTN = "a[href*='/applications/']"
     MY_GOODS_BTN = "a[href*='/goods/']"
-    ADD_A_GOOD_BTN = "a[href*='/goods/add/']"
     USERS_BTN = "a[href='/users/']"
     SITES_BTN = "a[href='/sites/']"
-    SITES_LINK = "a[href*='sites']"
     END_USER_ADVISORY_TILE = "a[href*='/end-users/']"
     RAISE_HMRC_QUERY_BTN = "a[href*='/raise-a-query/"
     USER_PROFILE_BTN = "a[href*='/users/profile/']"
