--- ui_automation_tests/pages/open_application/task_list.py
@@ -1,14 +1,0 @@
-from ui_automation_tests.shared.BasePage import BasePage
-from ui_automation_tests.shared.tools.helpers import scroll_to_element_by_id
-
-
-class OpenApplicationTaskListPage(BasePage):
-
-    LINK_COUNTRIES_ID = "link-countries"
-
-    def _click_link(self, element_id):
-        scroll_to_element_by_id(self.driver, element_id)
-        self.driver.find_element_by_id(element_id).click()
-
-    def click_countries_link(self):
-        self._click_link(self.LINK_COUNTRIES_ID)
