+++ ui_automation_tests/pages/sites_page.py
@@ -6,6 +6,3 @@ class SitesPage(BasePage):
 
     def click_sites_checkbox(self, no):
         self.driver.find_elements_by_css_selector(self.SITES_CHECKBOX)[no].click()
-
-    def get_checked_attribute_of_sites_checkbox(self, no):
-        return self.driver.find_elements_by_css_selector(self.SITES_CHECKBOX)[no].get_attribute("checked")
